task3 - refactoring package private classes and moving to controller
